rename: bottomSheetOption 파일명 변경, 가져오기 업데이트
